Second stage of D3D12 hooks separation
- finalized separation of D3D12 hook (removed everything D3D12 related from Overlay)
- activating separated D3D12 hook
- needs to call into scripts in some logical way (next step: create some ScriptingHandler or modify Scripting)
- needs to separate scripting things from overlay into some handler (dunno, step after above one)
- proper delta time calculation (+ possibly some statistics)
